+++ src/gt4py/next/program_processors/runners/dace_fieldview/__init__.py
@@ -13,7 +13,7 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 
-from gt4py.next.program_processors.runners.dace_fieldview.workflow import build_sdfg_from_gtir
+from gt4py.next.program_processors.runners.dace_fieldview.gtir_to_sdfg import build_sdfg_from_gtir
 
 
 __all__ = [
